Enhance Compiler Warnings Management and Refactor Code
This commit removes the `-Werror` flag from the C and C++ compiler options to prevent warnings from being treated as errors. It adds `-Wno-error` and several specific warning suppression options to improve code compilation flexibility. Additionally, it refactors code for clarity and correctness, including the correction of type handling in `obs-properties-widget.cpp` and adjustments in the `push-widget.cpp` file to improve memory management.

Modified:
  - `CMakeLists.txt`
  - `cmake/linux/compilerconfig.cmake`
  - `src/obs-multi-rtmp.cpp`
  - `src/obs-properties-widget.cpp`
  - `src/push-widget.cpp`

No files were added or removed in this commit.